4.19

+++ 4/controllers/blogs.js
@@ -1,23 +1,63 @@
 const blogsRouter = require('express').Router()
-const { replaceOne } = require('../models/blog')
+const jwt = require('jsonwebtoken')
+
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({})
+    .populate('user', { username: 1, name: 1 })
+
+  response.json(blogs)
+})
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
   response.json(blogs)
 })
 
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  } else {
+    return null
+  }
+};
+
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
+
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+
+
+  //const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  //if (!request.token || !decodedToken.id) {
+  
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'Token missing or invalid' })
+  }
 
   if (!body.title && !body.url) {
     response.status(400).json('Title and URL missing')
   }
 
-  const blog = new Blog(body)
+  const user = await User.findById(decodedToken.id)
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  })
 
   try {
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
   } catch(expection) {
     next(expection)
@@ -25,8 +65,21 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'Token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'User without right to remove' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
@@ -38,7 +91,7 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
   try {
     const newBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-    response.json(blog)
+    response.status(200).json(blog)
   } catch(expection) {
     next(expection)
   }
